Python_Essentials_for_MLOps/Project_2/requirements.txt
apache-airflow[pandas]
pytest==7.4.2
pytest-sugar==0.9.7
pylint==3.0.0
requests>=2.31
xmltodict==0.13.0
pendulum==2.1.2
